refactor: split tactical HUD armor renderer
